pyproject.toml
[project]
name = "benchbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Travior", email = "30752199+Travior@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "duckdb>=1.4.2",
    "instructor>=1.13.0",
    "litellm>=1.80.5",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "rich>=14.0.0",
]

[project.scripts]
bench = "benchbench.cli:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
